chore: Remove unused code files
