+++ backend/src/prisma/prisma.module.ts
@@ -1,7 +1,9 @@
-import { Module } from '@nestjs/common';
-import { PrismaService } from './prisma.service';
+import { Global, Module } from '@nestjs/common'
+import { PrismaService } from './prisma.service'
 
+@Global()
 @Module({
-  providers: [PrismaService]
+	providers: [PrismaService],
+	exports: [PrismaService]
 })
 export class PrismaModule {}
