+++ backend/src/user/dto/update-user.dto.ts
@@ -1,4 +1,4 @@
 import { PartialType } from '@nestjs/mapped-types';
-import { CreateAuthDto } from './create-auth.dto';
+import { CreateAuthDto } from './create-user.dto';
 
 export class UpdateAuthDto extends PartialType(CreateAuthDto) {}
