+++ backend/src/user/user.controller.ts
@@ -1,0 +1,34 @@
+import { Controller, Get, Post, Body, Patch, Param, Delete } from '@nestjs/common'
+import { AuthService } from './user.service'
+import { CreateAuthDto } from './dto/create-user.dto'
+import { UpdateAuthDto } from './dto/update-user.dto'
+
+@Controller('auth')
+export class AuthController {
+	constructor(private readonly authService: AuthService) {}
+
+	@Post()
+	create(@Body() createAuthDto: CreateAuthDto) {
+		return this.authService.create(createAuthDto)
+	}
+
+	@Get()
+	findAll() {
+		return this.authService.findAll()
+	}
+
+	@Get(':id')
+	findOne(@Param('id') id: string) {
+		return this.authService.findOne(+id)
+	}
+
+	@Patch(':id')
+	update(@Param('id') id: string, @Body() updateAuthDto: UpdateAuthDto) {
+		return this.authService.update(+id, updateAuthDto)
+	}
+
+	@Delete(':id')
+	remove(@Param('id') id: string) {
+		return this.authService.remove(+id)
+	}
+}
