+++ backend/src/user/user.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@nestjs/common';
-import { CreateAuthDto } from './dto/create-auth.dto';
-import { UpdateAuthDto } from './dto/update-auth.dto';
+import { CreateAuthDto } from './dto/create-user.dto';
+import { UpdateAuthDto } from './dto/update-user.dto';
 
 @Injectable()
 export class AuthService {
